requirements.txt
pytest>=7.4.0
allure-pytest>=2.13.5
python-dotenv>=1.0.1
pydantic>=2.6.1
Appium-Python-Client==2.10.0
requests>=2.31.0
